backend/requirements.txt
argon2-cffi==20.1.0
django==3.1.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.5.0
psycopg2-binary==2.8.6; sys_platform == 'linux'
psycopg2==2.8.6; sys_platform == 'win32'
serpy==0.3.1
gunicorn==20.0.4
Pillow==8.0.1
